Kaller mottak direkte med pub ingress. Fjerner referanse til apigw-apikey
