+++ nais/dev-gcp.json
@@ -11,7 +11,8 @@
   "externalHosts": [
     "oauth2.googleapis.com",
     "login.microsoftonline.com",
-    "k9-selvbetjening-oppslag.dev-fss-pub.nais.io"
+    "k9-selvbetjening-oppslag.dev-fss-pub.nais.io",
+    "omsorgspengerutbetalingsoknad-arbeidstaker-mottak.dev-fss-pub.nais.io"
   ],
   "env": {
     "REDIS_HOST": "omsorgspengerutbetalingsoknad-arbeidstaker-api-redis",
@@ -19,7 +20,7 @@
     "LOGIN_SERVICE_V1_DISCOVERY_ENDPOINT": "https://login.microsoftonline.com/navtestb2c.onmicrosoft.com/discovery/v2.0/.well-known/openid-configuration?p=b2c_1a_idporten_ver1",
     "COOKIE_NAME": "selvbetjening-idtoken",
     "K9_OPPSLAG_REGISTER_URL": "https://k9-selvbetjening-oppslag.dev-fss-pub.nais.io",
-    "OMSORGSPENGERUTBETALINGSOKNAD_MOTTAK_BASE_URL": "https://api-gw-q1.oera.no/helse-reverse-proxy/omsorgspengerutbetalingsoknad-arbeidstaker-mottak",
+    "OMSORGSPENGERUTBETALINGSOKNAD_MOTTAK_BASE_URL": "https://omsorgspengerutbetalingsoknad-arbeidstaker-mottak.dev-fss-pub.nais.io",
     "K9_MELLOMLAGRING_SERVICE_DISCOVERY": "http://k9-mellomlagring",
     "K9_MELLOMLAGRING_CLIENT_ID": "8d0460bd-ea81-4b08-b524-cf04874a794c/.default",
     "OMSORGSPENGERUTBETALINGSOKNAD_MOTTAK_CLIENT_ID": "8de5e926-fefe-47ed-bfb2-83645314c1d3/.default"
