+++ nais/prod-gcp.json
@@ -12,7 +12,8 @@
   "externalHosts": [
     "oauth2.googleapis.com",
     "login.microsoftonline.com",
-    "k9-selvbetjening-oppslag.prod-fss-pub.nais.io"
+    "k9-selvbetjening-oppslag.prod-fss-pub.nais.io",
+    "omsorgspengerutbetalingsoknad-arbeidstaker-mottak.prod-fss-pub.nais.io"
   ],
   "env": {
     "REDIS_HOST": "omsorgspengerutbetalingsoknad-arbeidstaker-api-redis",
@@ -20,7 +21,7 @@
     "LOGIN_SERVICE_V1_DISCOVERY_ENDPOINT": "https://login.microsoftonline.com/navnob2c.onmicrosoft.com/discovery/v2.0/.well-known/openid-configuration?p=b2c_1a_idporten",
     "COOKIE_NAME": "selvbetjening-idtoken",
     "K9_OPPSLAG_REGISTER_URL": "https://k9-selvbetjening-oppslag.prod-fss-pub.nais.io",
-    "OMSORGSPENGERUTBETALINGSOKNAD_MOTTAK_BASE_URL": "https://api-gw.oera.no/helse-reverse-proxy/omsorgspengerutbetalingsoknad-arbeidstaker-mottak",
+    "OMSORGSPENGERUTBETALINGSOKNAD_MOTTAK_BASE_URL": "https://omsorgspengerutbetalingsoknad-arbeidstaker-mottak.prod-fss-pub.nais.io",
     "K9_MELLOMLAGRING_SERVICE_DISCOVERY": "http://k9-mellomlagring",
     "K9_MELLOMLAGRING_CLIENT_ID": "19aaf0b2-f40a-4a64-bf7f-fd2dd62f0552/.default",
     "OMSORGSPENGERUTBETALINGSOKNAD_MOTTAK_CLIENT_ID": "90b18030-2c75-4fc4-8810-42e0b164503e/.default"
