--- src/main/kotlin/no/nav/omsorgspengerutbetaling/general/auth/ApiGatewayApiKey.kt
@@ -1,3 +1,0 @@
-package no.nav.omsorgspengerutbetaling.general.auth
-
-data class ApiGatewayApiKey(val value : String, val headerKey : String = "x-nav-apiKey")
