+++ src/main/kotlin/no/nav/omsorgspengerutbetaling/mottak/OmsorgpengesøknadMottakGateway.kt
@@ -16,7 +16,6 @@ import no.nav.helse.dusseldorf.ktor.metrics.Operation
 import no.nav.helse.dusseldorf.oauth2.client.AccessTokenClient
 import no.nav.helse.dusseldorf.oauth2.client.CachedAccessTokenClient
 import no.nav.omsorgspengerutbetaling.general.CallId
-import no.nav.omsorgspengerutbetaling.general.auth.ApiGatewayApiKey
 import no.nav.omsorgspengerutbetaling.soknad.KomplettSøknad
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -26,8 +25,7 @@ import java.net.URI
 class OmsorgpengesøknadMottakGateway(
     baseUrl: URI,
     private val accessTokenClient: AccessTokenClient,
-    private val sendeSoknadTilProsesseringScopes: Set<String>,
-    private val apiGatewayApiKey: ApiGatewayApiKey
+    private val sendeSoknadTilProsesseringScopes: Set<String>
 ) : HealthCheck {
 
     private companion object {
@@ -71,8 +69,7 @@ class OmsorgpengesøknadMottakGateway(
             .header(
                 HttpHeaders.ContentType to "application/json",
                 HttpHeaders.XCorrelationId to callId.value,
-                HttpHeaders.Authorization to authorizationHeader,
-                apiGatewayApiKey.headerKey to apiGatewayApiKey.value
+                HttpHeaders.Authorization to authorizationHeader
             )
 
         val (request, _, result) = Operation.monitored(
