+++ src/main/resources/application.conf
@@ -52,10 +52,6 @@ nav {
     }
   }
   authorization {
-    api_gateway {
-      api_key = ""
-      api_key = ${?APIGW_APIKEY}
-    }
     cookie_name = ""
     cookie_name = ${?COOKIE_NAME}
   }
